requirements.txt
# Core Flask and Database
flask
flask-sqlalchemy
flask-login
werkzeug
alembic

# PDF Processing Libraries
pypdf
pikepdf
reportlab
pdfplumber
pdf2image
pymupdf
pillow

# Document Conversion
python-docx
openpyxl

# Utilities
python-dotenv

# Authentication and Security
itsdangerous
flask-mail
twilio
redis

# AI/ML and Advanced Features
celery
langchain
transformers
torch
sentence-transformers
faiss-cpu
spacy
gensim
flask-restx
spacy
spacy

# Observability
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-celery
opentelemetry-exporter-jaeger

# ML Model Management
mlflow
scikit-learn